# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Documentation for Kopibang Coffee Point of Sales \u0026 Customer App.",
        "title": "Kopibang Coffee API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Sefza",
            "email": "admin@sefza.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/admin/orders": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates order, process redeem if any, returns receipt text and Earn QR",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Transaction"
                ],
                "summary": "Admin Input Order",
                "parameters": [
                    {
                        "description": "Order Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CreateOrderResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/admin/raw-materials": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetch all raw material stock lists",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Raw Material"
                ],
                "summary": "Get All Raw Materials",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.RawMaterialResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin can add new raw material stock",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Raw Material"
                ],
                "summary": "Add Raw Material",
                "parameters": [
                    {
                        "description": "Material Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RawMaterialRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/raw-materials/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin can update existing raw material details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Raw Material"
                ],
                "summary": "Update Raw Material",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Material ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Material Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RawMaterialRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin can delete a raw material",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Raw Material"
                ],
                "summary": "Delete Raw Material",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Material ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/settings/barista-status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin toggle to set available (online) or not available (offline)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Settings"
                ],
                "summary": "Update Barista Status",
                "parameters": [
                    {
                        "description": "Availability Status",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BaristaStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/points/redeem-qr": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "User generates QR token to show to Admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Transaction"
                ],
                "summary": "User Request Redeem QR",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.RedeemQRResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/points/scan-earn": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "User scans QR generated from POS to earn points",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Transaction"
                ],
                "summary": "User Scans Admin's QR",
                "parameters": [
                    {
                        "description": "Earn Token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ScanEarnPointRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.BaristaStatusRequest": {
            "type": "object",
            "required": [
                "is_available"
            ],
            "properties": {
                "is_available": {
                    "description": "Pakai pointer agar nilai false bisa tervalidasi",
                    "type": "boolean"
                }
            }
        },
...
        "dto.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.CreateOrderRequest": {
            "type": "object",
            "required": [
                "items"
            ],
            "properties": {
                "discount": {
                    "type": "integer",
                    "minimum": 0
                },
                "is_redeem": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/dto.OrderItemRequest"
                    }
                },
                "order_voucher_code": {
                    "type": "string"
                },
                "redeem_token": {
                    "description": "Wajib diisi jika IsRedeem true",
                    "type": "string"
                },
                "send_email_to": {
                    "description": "Opsional",
                    "type": "string"
                }
            }
        },
        "dto.CreateOrderResponse": {
            "type": "object",
            "properties": {
                "earn_point_token": {
                    "description": "QR Code untuk user scan",
                    "type": "string"
                },
                "final_amount": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "string"
                },
                "receipt_text": {
                    "description": "Teks untuk dilempar ke WA via Flutter",
                    "type": "string"
                }
            }
        },
...
        "dto.OrderItemRequest": {
            "type": "object",
            "required": [
                "product_id",
                "quantity"
            ],
            "properties": {
                "product_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "dto.RawMaterialRequest": {
            "type": "object",
            "required": [
                "name",
                "price",
                "quantity",
                "unit"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer",
                    "minimum": 0
                },
                "quantity": {
                    "type": "number",
                    "minimum": 0
                },
                "source": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "dto.RawMaterialResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                },
                "source": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "dto.RedeemQRResponse": {
            "type": "object",
            "properties": {
                "redeem_token": {
                    "type": "string"
                }
            }
        },
...
        "dto.ScanEarnPointRequest": {
            "type": "object",
            "required": [
                "earn_token"
            ],
            "properties": {
                "earn_token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by a space and your JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}